refactor(BrowserData-page): split extension component
